Added ugly outputs for code tracing
Conclusion: current issue is in SelectVictim - incorrect use of uniform distribution function and random number generator.

+++ src/selectvictim.h
@@ -9,7 +9,7 @@ class VictimGenerator {
 
     public:
 
-    int selectVictim(size_t bucket_size); // generates a random number from [0, bucket_size-1] to represent index of the victim
+    int selectVictim(int bucket_size); // generates a random number from [0, bucket_size-1] to represent index of the victim
 };
 
 #endif // SELECTVICTIM_H
